Profile Page as seen by others
Created the profile page of a user as seen by others users. Designed
the page.
Created the messaging GUI which pops up on the bottom right of the
screen and able to be closed by pressing “X”
Implemented the Friends button. When clicked all the friends will show.
Implemented the Reviews button. When clicked, the user’s rating by
other users and the comments left are shown.